Skip sgabioslog if VGA is detected.
The sgabioslog area is only read by the int 10h, function 08h handler
(read_attr_char), which is a NOP when VGA is attached. This skips both
the initialization and the writing of every character to the sgabioslog
buffer if a vga is attached. This saves us a possibly error-prone EBDA
allocation in that case.